+++ dev/dir-structure/CHANGELOG.md
@@ -1,3 +1,33 @@
+# Unreleased Changes
+
+## `ext_filter` macro
+
+The `ext_filter` macro has been introduced to simplify the creation of file extension filters:
+
+```rust
+use dir_structure::ext_filter;
+
+ext_filter!(RsSourceFileFilter, "rs");
+```
+
+## `DirChildren::get_mut`
+
+Mutable version of `DirChildren::get`, allowing you to get a mutable reference to a child by its name.
+
+## `DirChildren::iter_mut`
+
+A mutable iterator over the children of a `DirChildren` structure, allowing you to mutate the children while iterating.
+
+## `DeferredReadOrOwn::perform_and_store_read` now returns a mutable reference
+
+This method now returns a mutable reference to the read value, allowing you to modify it directly after reading, as opposed to an immutable reference, as was previously the case.
+
+## `Versioned::reset`
+
+Resets the versioned value to its initial version, marking it as clean.
+
+This is an unsafe API, as it allows you to reset the version without updating the files on the file system, so a further call to `write_to` will **not** write the value to disk, even if it was previously marked as dirty. **Use with caution.**
+
 # `0.1.5`
 
 Released: 2025-04-16